refactor(api-gateway): consolidate friends export adapters
Refs: #244
